refactor(uploadFile): componentize functionality
