+++ codebase/frontend/src/app/shared/components/UploadFile/index.ts
@@ -1,0 +1,1 @@
+export { default as UploadFile } from "./UploadFile";
