+++ codebase/frontend/src/app/shared/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./Breadcrumb";
 export * from "./ConfirmationDialog";
+export * from "./UploadFile";
